+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Accept `-` as a special path to read from `stdin`. If `-` is used multiple times, only the first one is effective.
 * Add `-N`/`--no-line-numbers` to hide line numbers
 * Add `--no-trailing-newlines` to remove the (new) trailing newline after each match
+* Add `--first-line-number` to show only each match’s first line number
 * Fix `--version` that was failing with an `IllegalArgumentException` (#1)
 
 ## 0.4.0 (2020/10/24)